# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BLACK DOG SECURITY API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SubmitRequestBody {
  name: string;
  organization: string;
  email: string;
  requestType: string;
  subject: string;
  message: string;
  linkedSystem?: string;
}

export interface BdsRequest {
  id: string;
  name: string;
  organization: string;
  email: string;
  requestType: string;
  subject: string;
  message: string;
  linkedSystem: string;
  status: string;
  /** @nullable */
  internalNotes?: string | null;
  packetSent: boolean;
  /** @nullable */
  packetSentAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ActionHistoryEntry {
  id: number;
  requestId: string;
  action: string;
  actor: string;
  /** @nullable */
  note?: string | null;
  timestamp: string;
}

export interface BdsRequestDetail {
  id: string;
  name: string;
  organization: string;
  email: string;
  requestType: string;
  subject: string;
  message: string;
  linkedSystem: string;
  status: string;
  /** @nullable */
  internalNotes?: string | null;
  packetSent: boolean;
  /** @nullable */
  packetSentAt?: string | null;
  createdAt: string;
  updatedAt: string;
  history: ActionHistoryEntry[];
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface LoginResponse {
  token: string;
  authenticated: boolean;
}

export interface SessionStatus {
  authenticated: boolean;
}

export interface UpdateStatusBody {
  status: string;
  note?: string;
}

export interface UpdateNotesBody {
  internalNotes: string;
}

export interface SystemLogEntry {
  id: number;
  event: string;
  /** @nullable */
  details?: string | null;
  timestamp: string;
}

export type ListCommanderRequestsParams = {
  status?: string;
  q?: string;
};