# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BLACK DOG SECURITY API
 * OpenAPI spec version: 0.2.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Submit an access/briefing request
 */
export const SubmitRequestBody = zod.object({
  name: zod.string(),
  organization: zod.string(),
  email: zod.string(),
  requestType: zod.string(),
  subject: zod.string(),
  message: zod.string(),
  linkedSystem: zod.string().optional(),
});

/**
 * @summary Admin login
 */
export const CommanderLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const CommanderLoginResponse = zod.object({
  token: zod.string(),
  authenticated: zod.boolean(),
});

/**
 * @summary Admin logout
 */
export const CommanderLogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Check session status
 */
export const GetCommanderSessionResponse = zod.object({
  authenticated: zod.boolean(),
});

/**
 * @summary List all requests
 */
export const ListCommanderRequestsQueryParams = zod.object({
  status: zod.coerce.string().optional(),
  q: zod.coerce.string().optional(),
});

export const ListCommanderRequestsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  organization: zod.string(),
  email: zod.string(),
  requestType: zod.string(),
  subject: zod.string(),
  message: zod.string(),
  linkedSystem: zod.string(),
  status: zod.string(),
  internalNotes: zod.string().nullish(),
  packetSent: zod.boolean(),
  packetSentAt: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListCommanderRequestsResponse = zod.array(
  ListCommanderRequestsResponseItem,
);

/**
 * @summary Get request with action history
 */
export const GetCommanderRequestParams = zod.object({
  id: zod.coerce.string(),
});

export const GetCommanderRequestResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  organization: zod.string(),
  email: zod.string(),
  requestType: zod.string(),
  subject: zod.string(),
  message: zod.string(),
  linkedSystem: zod.string(),
  status: zod.string(),
  internalNotes: zod.string().nullish(),
  packetSent: zod.boolean(),
  packetSentAt: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  history: zod.array(
    zod.object({
      id: zod.number(),
      requestId: zod.string(),
      action: zod.string(),
      actor: zod.string(),
      note: zod.string().nullish(),
      timestamp: zod.string(),
    }),
  ),
});

/**
 * @summary Update request status
 */
export const UpdateRequestStatusParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateRequestStatusBody = zod.object({
  status: zod.string(),
  note: zod.string().optional(),
});

export const UpdateRequestStatusResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  organization: zod.string(),
  email: zod.string(),
  requestType: zod.string(),
  subject: zod.string(),
  message: zod.string(),
  linkedSystem: zod.string(),
  status: zod.string(),
  internalNotes: zod.string().nullish(),
  packetSent: zod.boolean(),
  packetSentAt: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update internal notes on a request
 */
export const UpdateRequestNotesParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateRequestNotesBody = zod.object({
  internalNotes: zod.string(),
});

export const UpdateRequestNotesResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  organization: zod.string(),
  email: zod.string(),
  requestType: zod.string(),
  subject: zod.string(),
  message: zod.string(),
  linkedSystem: zod.string(),
  status: zod.string(),
  internalNotes: zod.string().nullish(),
  packetSent: zod.boolean(),
  packetSentAt: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Dispatch information packet to requester
 */
export const DispatchPacketParams = zod.object({
  id: zod.coerce.string(),
});

export const DispatchPacketResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  organization: zod.string(),
  email: zod.string(),
  requestType: zod.string(),
  subject: zod.string(),
  message: zod.string(),
  linkedSystem: zod.string(),
  status: zod.string(),
  internalNotes: zod.string().nullish(),
  packetSent: zod.boolean(),
  packetSentAt: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get system event log
 */
export const GetSystemLogsResponseItem = zod.object({
  id: zod.number(),
  event: zod.string(),
  details: zod.string().nullish(),
  timestamp: zod.string(),
});
export const GetSystemLogsResponse = zod.array(GetSystemLogsResponseItem);